requirements.txt
requests>=2.31.0
requests-toolbelt>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0